+++ packages/components/src/Form/Fieldset/Fieldset.tsx
@@ -24,7 +24,7 @@
 
  */
 
-import React, { forwardRef, ReactNode, Ref } from 'react'
+import React, { forwardRef, ReactNode, Ref, createContext } from 'react'
 import styled from 'styled-components'
 import {
   CompatibleHTMLProps,
@@ -65,6 +65,11 @@ export interface FieldsetProps
    * @default 'inline ? 'medium' : 'small'
    */
   gap?: SpacingSizes
+  /*
+   * Hide labels on Fields within this Fieldset=
+   * @default false
+   */
+  labelsCollapse?: boolean
 }
 
 const accordionIndicatorDefaults: AccordionIndicatorProps = {
@@ -77,6 +82,12 @@ const accordionIndicatorDefaults: AccordionIndicatorProps = {
   indicatorSize: 'medium',
 }
 
+export interface FieldsetContext {
+  labelsCollapse?: boolean
+}
+
+export const FieldsetContext = createContext<FieldsetContext>({})
+
 const FieldsetLayout = forwardRef(
   (props: FieldsetProps, ref: Ref<HTMLDivElement>) => {
     const {
@@ -85,6 +96,7 @@ const FieldsetLayout = forwardRef(
       gap,
       legend,
       children,
+      labelsCollapse,
       ...restProps
     } = omit(props, [...AccordionControlPropKeys])
 
@@ -138,7 +150,17 @@ const FieldsetLayout = forwardRef(
       content
     )
 
+<<<<<<< HEAD
     return <div {...omitStyledProps(restProps)}>{renderedFieldset}</div>
+=======
+    return (
+      <FieldsetContext.Provider
+        value={{ labelsCollapse: labelsCollapse || false }}
+      >
+        <div className={className}>{renderedFieldset}</div>
+      </FieldsetContext.Provider>
+    )
+>>>>>>> labelsCollapse as prop and using context
   }
 )